# apps/luna-shopper-backend/gateway/docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/zones/{zoneId}/merges": {
      "post": {
        "operationId": "MergeController_request_v1",
        "parameters": [
          {
            "name": "zoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestMergeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The `merge.request` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/merge.MergeRequestView"
                }
              }
            }
          },
          "400": {
            "description": "`validation_failed` — Some of the submitted values are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`conflict` — That request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "merges"
        ]
      },
      "get": {
        "operationId": "MergeController_list_v1",
        "parameters": [
          {
            "name": "zoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The `merge.list` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/merge.MergeRequestPage"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "merges"
        ]
      }
    },
    "/v1/merges/{id}/approve": {
      "post": {
        "operationId": "MergeController_approve_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The `merge.approve` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/merge.MergeRequestView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`conflict` — That request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "merges"
        ]
      }
    },
    "/v1/merges/{id}/reject": {
      "post": {
        "operationId": "MergeController_reject_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The `merge.reject` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/merge.MergeRequestView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`conflict` — That request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "merges"
        ]
      }
    },
    "/v1/merges/{id}/cancel": {
      "post": {
        "operationId": "MergeController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The `merge.cancel` contract response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/merge.MergeRequestView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — You need to sign in to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — You do not have permission to do that.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — The requested resource was not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "`conflict` — That request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — Too many requests. Please slow down and try again shortly.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — Something went wrong on our side. Please try again.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "merges"
        ]
      }
    },
...
  },
  "info": {
    "title": "Luna Shopper API",
    "description": "Public API for the Luna Shopper shared shopping lists. Response schemas are projected from the same JSON Schemas the services validate their broker messages against, so what is documented here is what the wire carries.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RequestMergeDto": {
        "type": "object",
        "properties": {
          "sourceUserId": {
            "type": "string",
            "format": "uuid",
            "description": "Account whose data is taken from"
          },
          "targetUserId": {
            "type": "string",
            "format": "uuid",
            "description": "Account the data is moved into"
          }
        },
        "required": [
          "sourceUserId",
          "targetUserId"
        ]
      },
...
      "enums.MergeRequestStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "CANCELLED"
        ]
      },
...
      "merge.MergeRequestPage": {
        "type": "object",
        "description": "A cursor paginated page. `nextCursor` is null on the last page; otherwise pass it back as the `cursor` query parameter to fetch the next one.",
        "additionalProperties": false,
        "required": [
          "items",
          "nextCursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/merge.MergeRequestView"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "merge.MergeRequestView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "zoneId": {
            "type": "string",
            "minLength": 1
          },
          "sourceUserId": {
            "type": "string",
            "minLength": 1
          },
          "targetUserId": {
            "type": "string",
            "minLength": 1
          },
          "requestedByUserId": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/enums.MergeRequestStatus"
          },
          "resolvedByUserId": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "zoneId",
          "sourceUserId",
          "targetUserId",
          "requestedByUserId",
          "status",
          "resolvedByUserId"
        ],
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "description": "The RFC 7807 error envelope, served as `application/problem+json`. `message` is already translated to the request locale, so a client can show it without knowing any backend error code; `code` is the stable value to branch on.",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "message",
          "correlationId"
        ],
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI identifying the problem type."
          },
          "title": {
            "type": "string",
            "description": "A short, stable problem summary."
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code, repeated in the body."
          },
          "code": {
            "type": "string",
            "enum": [
              "validation_failed",
              "unauthorized",
              "forbidden",
              "not_found",
              "conflict",
              "rate_limited",
              "internal"
            ],
            "description": "The stable machine code the client branches on."
          },
          "detail": {
            "type": "string",
            "description": "Extra context about this occurrence, when there is any."
          },
          "message": {
            "type": "string",
            "description": "The message to show the user, in the request's locale."
          },
          "correlationId": {
            "type": "string",
            "description": "Identifies this request across every service log; quote it in a bug report."
          },
          "errors": {
            "type": "object",
            "description": "Per field validation messages. Present only when `code` is `validation_failed`.",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}